requirements.txt
requests
transformers
python-dotenv
schedule
